# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Typewriter API
  version: 1.0.0
  description: API für Typewriter-Sessions und LiveSessions (Streaming via WebSocket)
servers:
  - url: https://example.com
    description: Platzhalter
  - url: http://localhost:{port}
    description: Lokales Development (Port = ENV PORT oder Default 3001)
    variables:
      port:
        default: "3001"
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required: [success, error, message]
...
    Note:
      type: object
      description: Offsets gelten nur innerhalb dieser Session-Version.
      properties:
        id:
          type: integer
        session_id:
          type: integer
        start_offset:
          type: integer
          minimum: 0
        end_offset:
          type: integer
          minimum: 1
        note:
          type: string
      required: [id, session_id, start_offset, end_offset, note]
paths:
...
  /api/v1/notes/{note_id}:
    patch:
      summary: Notiz aktualisieren
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_offset:
                  type: integer
                  minimum: 0
                end_offset:
                  type: integer
                  minimum: 1
                note:
                  type: string
              description: Mindestens ein Feld muss gesetzt sein.
      responses:
        "200":
          description: Notiz aktualisiert
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Note'
                required: [success, data]
        "400":
          description: validation_error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: note_not_found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Notiz loeschen
      security:
        - apiKeyAuth: []
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Notiz geloescht
        "404":
          description: note_not_found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'